refactor(iota-types): remove TransactionEffectsV2
