+++ crates/iota-types/src/full_checkpoint_content.rs
@@ -4,7 +4,6 @@
 
 use std::collections::BTreeMap;
 
-use itertools::Either;
 use serde::{Deserialize, Serialize};
 use tap::Pipe;
 
@@ -95,25 +94,8 @@ impl CheckpointTransaction {
     pub fn removed_objects_pre_version(&self) -> impl Iterator<Item = &Object> {
         // Iterator over id and versions for all deleted or wrapped objects
         match &self.effects {
-            TransactionEffects::V1(v1) => Either::Left(
-                // Effects v1 has deleted and wrapped objects versions as the "new" version, not
-                // the old one that was actually removed. So we need to take these
-                // and then look them up in the `modified_at_versions`.
-                // No need to chain unwrapped_then_deleted because these objects must have been
-                // wrapped before the transaction, hence they will not be in
-                // modified_at_versions / input_objects.
-                v1.deleted().iter().chain(v1.wrapped()).map(|(id, _, _)| {
-                    // lookup the old version for mutated objects
-                    let (_, old_version) = v1
-                        .modified_at_versions()
-                        .iter()
-                        .find(|(oid, _old_version)| oid == id)
-                        .expect("deleted/wrapped object must have entry in 'modified_at_versions'");
-                    (id, old_version)
-                }),
-            ),
-            TransactionEffects::V2(v2) => {
-                Either::Right(v2.changed_objects().iter().filter_map(|(id, change)| {
+            TransactionEffects::V1(v1) => {
+                v1.changed_objects().iter().filter_map(|(id, change)| {
                     match (
                         &change.input_state,
                         &change.output_state,
@@ -134,7 +116,7 @@ impl CheckpointTransaction {
                         ) => Some((id, version)),
                         _ => None,
                     }
-                }))
+                })
             }
         }
         // Use id and version to lookup in input Objects
@@ -156,24 +138,8 @@ impl CheckpointTransaction {
     pub fn changed_objects(&self) -> impl Iterator<Item = (&Object, Option<&Object>)> {
         // Iterator over ((ObjectId, new version), Option<old version>)
         match &self.effects {
-            TransactionEffects::V1(v1) => Either::Left(
-                v1.created()
-                    .iter()
-                    .chain(v1.unwrapped())
-                    .map(|((id, version, _), _)| ((id, version), None))
-                    .chain(v1.mutated().iter().map(|((id, version, _), _)| {
-                        // lookup the old version for mutated objects
-                        let (_, old_version) = v1
-                            .modified_at_versions()
-                            .iter()
-                            .find(|(oid, _old_version)| oid == id)
-                            .expect("mutated object must have entry in modified_at_versions");
-
-                        ((id, version), Some(old_version))
-                    })),
-            ),
-            TransactionEffects::V2(v2) => {
-                Either::Right(v2.changed_objects().iter().filter_map(|(id, change)| {
+            TransactionEffects::V1(v1) => {
+                v1.changed_objects().iter().filter_map(|(id, change)| {
                     match (
                         &change.input_state,
                         &change.output_state,
@@ -181,7 +147,7 @@ impl CheckpointTransaction {
                     ) {
                         // Created Objects
                         (ObjectIn::NotExist, ObjectOut::ObjectWrite(_), IDOperation::Created) => {
-                            Some(((id, &v2.lamport_version), None))
+                            Some(((id, &v1.lamport_version), None))
                         }
                         (
                             ObjectIn::NotExist,
@@ -191,12 +157,12 @@ impl CheckpointTransaction {
 
                         // Unwrapped Objects
                         (ObjectIn::NotExist, ObjectOut::ObjectWrite(_), IDOperation::None) => {
-                            Some(((id, &v2.lamport_version), None))
+                            Some(((id, &v1.lamport_version), None))
                         }
 
                         // Mutated Objects
                         (ObjectIn::Exist(((old_version, _), _)), ObjectOut::ObjectWrite(_), _) => {
-                            Some(((id, &v2.lamport_version), Some(old_version)))
+                            Some(((id, &v1.lamport_version), Some(old_version)))
                         }
                         (
                             ObjectIn::Exist(((old_version, _), _)),
@@ -206,7 +172,7 @@ impl CheckpointTransaction {
 
                         _ => None,
                     }
-                }))
+                })
             }
         }
         // Lookup Objects in output Objects as well as old versions for mutated objects
@@ -231,13 +197,8 @@ impl CheckpointTransaction {
     pub fn created_objects(&self) -> impl Iterator<Item = &Object> {
         // Iterator over (ObjectId, version) for created objects
         match &self.effects {
-            TransactionEffects::V1(v1) => Either::Left(
-                v1.created()
-                    .iter()
-                    .map(|((id, version, _), _)| (id, version)),
-            ),
-            TransactionEffects::V2(v2) => {
-                Either::Right(v2.changed_objects().iter().filter_map(|(id, change)| {
+            TransactionEffects::V1(v1) => {
+                v1.changed_objects().iter().filter_map(|(id, change)| {
                     match (
                         &change.input_state,
                         &change.output_state,
@@ -245,7 +206,7 @@ impl CheckpointTransaction {
                     ) {
                         // Created Objects
                         (ObjectIn::NotExist, ObjectOut::ObjectWrite(_), IDOperation::Created) => {
-                            Some((id, &v2.lamport_version))
+                            Some((id, &v1.lamport_version))
                         }
                         (
                             ObjectIn::NotExist,
@@ -255,7 +216,7 @@ impl CheckpointTransaction {
 
                         _ => None,
                     }
-                }))
+                })
             }
         }
         // Lookup Objects in output Objects as well as old versions for mutated objects
